requirements.txt
torch
ccxt
pandas
numpy
scikit-learn
requests
ta
openai
fastapi
uvicorn
python-dotenv